moved SDC related files to their own folder
